refactor: 옵션 도메인 규칙 엔티티로 옮김
