+++ src/main/java/org/son/sonstudy/domain/product/model/ProductOption.java
@@ -3,6 +3,8 @@ package org.son.sonstudy.domain.product.model;
 import io.hypersistence.utils.hibernate.id.Tsid;
 import jakarta.persistence.*;
 import lombok.*;
+import org.son.sonstudy.common.api.code.ErrorCode;
+import org.son.sonstudy.common.exception.CustomException;
 
 @Entity
 @Getter
@@ -32,9 +34,31 @@ public class ProductOption {
 
     @Builder
     public ProductOption(int size, int cost, int stock) {
+        validateSize(size);
+        validateCost(cost);
+        validateStock(stock);
+
         this.size = size;
         this.cost = cost;
         this.stock = stock;
         this.totalSales = 0L;
+    }
+
+    private void validateSize(int size) {
+        if (size < 100) {
+            throw new CustomException(ErrorCode.INVALID_PRODUCT_SIZE);
+        }
+    }
+
+    private void validateCost(int cost) {
+        if (cost < 0) {
+            throw new CustomException(ErrorCode.INVALID_PRODUCT_COST);
+        }
+    }
+
+    private void validateStock(int stock) {
+        if (stock < 0) {
+            throw new CustomException(ErrorCode.INVALID_STOCK);
+        }
     }
 }
